refactor: move functions for work with answers matrix to own module manage_matrix.js
