+++ nonograms/scripts/game_logic.js
@@ -158,26 +158,6 @@ function manageCell(targetCell) {
   setAnswer(sign, targetCellCol, targetCellRow);
 }
 
-function addDataToMatrix(sign, targetCellCol, targetCellRow) {
-  if (sign.includes('fill')) {
-    vars.answerCells.col_answer[targetCellCol][targetCellRow] += 1;
-    vars.answerCells.row_answer[targetCellRow][targetCellCol] += 1;
-  } else if (sign.includes('cross')) {
-    vars.answerCells.col_cross[targetCellCol][targetCellRow] += 1;
-    vars.answerCells.row_cross[targetCellRow][targetCellCol] += 1;
-  }
-}
-
-function eraseDataFromMatrix(sign, targetCellCol, targetCellRow) {
-  if (sign.includes('fill')) {
-    vars.answerCells.col_answer[targetCellCol][targetCellRow] -= 1;
-    vars.answerCells.row_answer[targetCellRow][targetCellCol] -= 1;
-  } else if (sign.includes('cross')) {
-    vars.answerCells.col_cross[targetCellCol][targetCellRow] -= 1;
-    vars.answerCells.row_cross[targetCellRow][targetCellCol] -= 1;
-  }
-}
-
 function showPopUp() {
   timer.stopTimer(true);
   setBlock(true);
